requirements.txt
Flask
Flask-SQLAlchemy
cx_Oracle
python-dotenv
